requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.0
python-dotenv==1.0.1
twilio==9.3.0
google-cloud-speech==2.27.0
google-cloud-vision==3.7.4
pyairtable==2.3.3
python-multipart==0.0.9
